Introduce shadow resources for NeutronCommands
In the process of adding the LBaaS v2 object model, we have the need to
support both v1 and v2. Naturally, what this means is that we will need
to handle both the v1 objects as well as the v2 object simultaneously,
despite the fact that they have different attributes.

From this client's persepective, there is currently no good way to do
so side-by-side. Currently, the code interpolates the command to be
executed based upon the resource being acted against.  For instance,
'lb-vip-create' would be interpolated as 'create_%s' % 'vip'

This is obviously a problem as there is no way for the client to know
which version of the resource to CRUD against. So, add the concept of
a shadow_resource to handle the case where the resource name is not
the same as the command to be run.

Thihs change also moves resource shadow_resource attributes into
NeutronCommand, as they were common to all the commands already.

Change-Id: Ib6d4c66dfa14574062323704d7ca135cb3887020